+++ spacepackets/ccsds/time.py
@@ -2,8 +2,10 @@ from __future__ import annotations
 import datetime
 import math
 import enum
+import struct
 import time
-from abc import abstractmethod
+from abc import abstractmethod, ABC
+from typing import Optional
 
 DAYS_CCSDS_TO_UNIX = -4383
 SECONDS_PER_DAY = 86400
@@ -45,22 +47,43 @@ def read_p_field(p_field: int) -> CcsdsTimeCodeId:
     return CcsdsTimeCodeId((p_field & 0x70) >> 4)
 
 
-class CcsdsTimeCode:
+class CcsdsTimeProvider(ABC):
+    @property
+    @abstractmethod
+    def pfield(self) -> bytes:
+        pass
+
+    @property
+    @abstractmethod
+    def len(self) -> int:
+        pass
+
     @abstractmethod
     def pack(self) -> bytearray:
-        self.ccsds_id = CcsdsTimeCodeId.NONE
-        return bytearray()
+        pass
 
     @abstractmethod
-    def return_unix_seconds(self) -> int:
-        return 0
+    def read_from_raw(self, timestamp: bytes):
+        pass
 
     @abstractmethod
-    def return_time_string(self) -> str:
-        return ""
+    def as_unix_seconds(self) -> float:
+        pass
+
+    @abstractmethod
+    def as_datetime(self) -> datetime:
+        pass
+
+    def as_time_string(self) -> str:
+        return self.as_datetime().strftime("%Y-%m-%d %H:%M:%S.%f")
+
+    def ccsds_time_code(self) -> int:
+        if self.pfield == bytes():
+            return 0
+        return (self.pfield[0] >> 4) & 0b111
 
 
-class CdsShortTimestamp(CcsdsTimeCode):
+class CdsShortTimestamp(CcsdsTimeProvider):
     """Unpacks the time datafield of the TM packet. Right now, CDS Short timeformat is used,
     and the size of the time stamp is expected to be seven bytes.
     """
@@ -69,88 +92,108 @@ class CdsShortTimestamp(CcsdsTimeCode):
     TIMESTAMP_SIZE = 7
 
     def __init__(self, ccsds_days: int, ms_of_day: int):
-        self.ccsds_id = CcsdsTimeCodeId.CDS
-        self.p_field = CdsShortTimestamp.CDS_SHORT_ID << 4
+        self._setup(ccsds_days, ms_of_day)
+
+    def _setup(self, ccsds_days: int, ms_of_day: int):
+        self.__p_field = bytes([CdsShortTimestamp.CDS_SHORT_ID << 4])
         # CCSDS recommends a 1958 Januar 1 epoch, which is different from the Unix epoch
-        self.ccsds_days = ccsds_days
-        self.unix_days = convert_ccsds_days_to_unix_days(self.ccsds_days)
-        self.unix_seconds = self.unix_days * (24 * 60 * 60)
-        self.ms_of_day = ms_of_day
-        self.seconds_of_day = self.ms_of_day / 1000.0
-        self.unix_seconds += self.seconds_of_day
-        if self.unix_seconds < 0:
-            date = datetime.datetime(1970, 1, 1) + datetime.timedelta(
-                seconds=self.unix_seconds
+        self._ccsds_days = ccsds_days
+        self._unix_seconds = 0
+        self._ms_of_day = ms_of_day
+        self._calculate_unix_seconds()
+        self._calculate_date_time()
+
+    def _calculate_unix_seconds(self):
+        unix_days = convert_ccsds_days_to_unix_days(self._ccsds_days)
+        self._unix_seconds = unix_days * (24 * 60 * 60)
+        seconds_of_day = self._ms_of_day / 1000.0
+        if self._unix_seconds < 0:
+            self._unix_seconds -= seconds_of_day
+        else:
+            self._unix_seconds += seconds_of_day
+
+    def _calculate_date_time(self):
+        if self._unix_seconds < 0:
+            self._date_time = datetime.datetime(1970, 1, 1) + datetime.timedelta(
+                seconds=self._unix_seconds
             )
         else:
-            date = datetime.datetime.utcfromtimestamp(self.unix_seconds)
-        self.time_string = date.strftime("%Y-%m-%d %H:%M:%S.%f")
+            self._date_time = datetime.datetime.utcfromtimestamp(self._unix_seconds)
+
+    @property
+    def pfield(self) -> bytes:
+        return self.__p_field
+
+    @property
+    def len(self) -> int:
+        return CdsShortTimestamp.TIMESTAMP_SIZE
+
+    def pack(self) -> bytearray:
+        cds_packet = bytearray()
+        cds_packet.extend(self.__p_field)
+        cds_packet.extend(struct.pack("!H", self._ccsds_days))
+        cds_packet.extend(struct.pack("!I", self._ms_of_day))
+        return cds_packet
 
     @classmethod
-    def init_from_unix_days(cls, unix_days: int, ms_of_day: int) -> CdsShortTimestamp:
+    def from_unix_days(cls, unix_days: int, ms_of_day: int) -> CdsShortTimestamp:
         return cls(
             ccsds_days=convert_unix_days_to_ccsds_days(unix_days=unix_days),
             ms_of_day=ms_of_day,
         )
 
-    def len(self):
-        return self.TIMESTAMP_SIZE
-
     @classmethod
-    def __empty(cls):
+    def empty(cls):
         return cls(ccsds_days=0, ms_of_day=0)
-
-    def pack(self) -> bytearray:
-        cds_packet = bytearray()
-        cds_packet.append(self.p_field)
-        cds_packet.append((self.ccsds_days & 0xFF00) >> 8)
-        cds_packet.append(self.ccsds_days & 0xFF)
-        cds_packet.append((self.ms_of_day & 0xFF000000) >> 24)
-        cds_packet.append((self.ms_of_day & 0x00FF0000) >> 16)
-        cds_packet.append((self.ms_of_day & 0x0000FF00) >> 8)
-        cds_packet.append(self.ms_of_day & 0x000000FF)
-        return cds_packet
 
     @classmethod
     def unpack(cls, time_field: bytes) -> CdsShortTimestamp:
-        if len(time_field) < cls.TIMESTAMP_SIZE:
-            raise ValueError
-        # TODO: check ID?
-        p_field = time_field[0]
-        ccsds_days = (time_field[1] << 8) | (time_field[2])
-        ms_of_day = (
-            (time_field[3] << 24)
-            | (time_field[4] << 16)
-            | (time_field[5]) << 8
-            | time_field[6]
-        )
+        ccsds_days, ms_of_day = CdsShortTimestamp.unpack_from_raw(time_field)
         return cls(ccsds_days=ccsds_days, ms_of_day=ms_of_day)
+
+    def read_from_raw(self, timestamp: bytes):
+        ccsds_days, ms_of_day = CdsShortTimestamp.unpack_from_raw(timestamp)
+        self._setup(ccsds_days, ms_of_day)
+
+    @staticmethod
+    def unpack_from_raw(raw: bytes) -> (int, int):
+        p_field = raw[0]
+        if (p_field >> 4) & 0b111 != CcsdsTimeCodeId.CDS:
+            raise ValueError(
+                f"Invalid CCSDS Time Code {p_field}, expected {CcsdsTimeCodeId.CDS}"
+            )
+        ccsds_days = (raw[1] << 8) | (raw[2])
+        ms_of_day = (raw[3] << 24) | (raw[4] << 16) | (raw[5]) << 8 | raw[6]
+        return ccsds_days, ms_of_day
 
     def __repr__(self):
         return (
-            f"{self.__class__.__name__}(ccsds_days={self.ccsds_days!r}, "
-            f"ms_of_day={self.ms_of_day!r})"
+            f"{self.__class__.__name__}(ccsds_days={self._ccsds_days!r}, "
+            f"ms_of_day={self._ms_of_day!r})"
         )
 
     def __str__(self):
-        return f"Date: {self.time_string} with representation {self!r}"
+        return f"Date {self._date_time!r} with representation {self!r}"
+
+    @classmethod
+    def from_current_time(cls) -> CdsShortTimestamp:
+        """Returns a seven byte CDS short timestamp with the current time"""
+        return cls.from_unix_days(
+            unix_days=(datetime.datetime.utcnow() - UNIX_EPOCH).days,
+            ms_of_day=cls.ms_of_day(),
+        )
 
     @staticmethod
-    def init_from_current_time() -> CdsShortTimestamp:
-        """Returns a seven byte CDS short timestamp with the current time"""
-        unix_days = (datetime.datetime.utcnow() - UNIX_EPOCH).days
-        seconds = time.time()
-        fraction_ms = seconds - math.floor(seconds)
-        days_ms = int((seconds % SECONDS_PER_DAY) * 1000 + fraction_ms)
-        time_packet = CdsShortTimestamp.init_from_unix_days(
-            unix_days=unix_days, ms_of_day=days_ms
+    def ms_of_day(seconds_since_epoch: Optional[float] = None):
+        if seconds_since_epoch is None:
+            seconds_since_epoch = time.time()
+        fraction_ms = seconds_since_epoch - math.floor(seconds_since_epoch)
+        return int(
+            math.floor((seconds_since_epoch % SECONDS_PER_DAY) * 1000 + fraction_ms)
         )
-        return time_packet
 
-    @abstractmethod
-    def return_unix_seconds(self) -> int:
-        return self.unix_seconds
+    def as_unix_seconds(self) -> float:
+        return self._unix_seconds
 
-    @abstractmethod
-    def return_time_string(self) -> str:
-        return self.time_string
+    def as_datetime(self) -> datetime:
+        return self._date_time