+++ spacepackets/ecss/pus_17_test.py
@@ -1,7 +1,10 @@
 from __future__ import annotations
 import enum
+from typing import Optional
+
+from spacepackets.ccsds.time import CcsdsTimeProvider
 from spacepackets.ecss.defs import PusServices
-from spacepackets.ecss.tm import CdsShortTimestamp, PusVersion, PusTelemetry
+from spacepackets.ecss.tm import PusVersion, PusTelemetry, AbstractPusTm
 
 
 class Subservices(enum.IntEnum):
@@ -9,32 +12,41 @@ class Subservices(enum.IntEnum):
     TM_REPLY = 2
 
 
-class Service17Tm:
+class Service17Tm(AbstractPusTm):
     def __init__(
         self,
         subservice: int,
-        time: CdsShortTimestamp = None,
+        time_provider: Optional[CcsdsTimeProvider] = None,
         ssc: int = 0,
         source_data: bytearray = bytearray([]),
         apid: int = -1,
         packet_version: int = 0b000,
-        pus_version: PusVersion = PusVersion.GLOBAL_CONFIG,
-        secondary_header_flag: bool = True,
         space_time_ref: int = 0b0000,
         destination_id: int = 0,
     ):
         self.pus_tm = PusTelemetry(
             service=PusServices.S17_TEST,
             subservice=subservice,
-            time=time,
+            time_provider=time_provider,
             seq_count=ssc,
             source_data=source_data,
             apid=apid,
             packet_version=packet_version,
-            secondary_header_flag=secondary_header_flag,
             space_time_ref=space_time_ref,
             destination_id=destination_id,
         )
+
+    @property
+    def service(self) -> int:
+        return self.pus_tm.service
+
+    @property
+    def subservice(self) -> int:
+        return self.pus_tm.subservice
+
+    @property
+    def source_data(self) -> bytes:
+        return self.pus_tm.source_data
 
     def pack(self) -> bytearray:
         return self.pus_tm.pack()