+++ spacepackets/ecss/tc.py
@@ -138,8 +138,6 @@ class PusTelecommand:
         """
         if apid == -1:
             apid = get_default_tc_apid()
-        secondary_header_flag = 1
-        logger = get_console_logger()
         self.pus_tc_sec_header = PusTcDataFieldHeader(
             service=service,
             subservice=subservice,
@@ -152,7 +150,7 @@ class PusTelecommand:
         )
         self.sp_header = SpacePacketHeader(
             apid=apid,
-            sec_header_flag=bool(secondary_header_flag),
+            sec_header_flag=True,
             packet_type=PacketTypes.TC,
             seq_flags=seq_flags,
             data_len=data_length,
@@ -337,6 +335,14 @@ class PusTelecommand:
     @property
     def subservice(self) -> int:
         return self.pus_tc_sec_header.subservice
+
+    @property
+    def source_id(self) -> int:
+        return self.pus_tc_sec_header.source_id
+
+    @source_id.setter
+    def source_id(self, source_id: int):
+        self.pus_tc_sec_header.source_id = source_id
 
     @property
     def seq_count(self) -> int: