+++ spacepackets/ecss/tm.py
@@ -3,6 +3,7 @@
 """
 from __future__ import annotations
 
+from abc import abstractmethod, ABC
 import struct
 from typing import Optional
 
@@ -17,7 +18,7 @@ from spacepackets.ccsds.spacepacket import (
     PacketTypes,
     SpacePacket,
 )
-from spacepackets.ccsds.time import CdsShortTimestamp, read_p_field
+from spacepackets.ccsds.time import CdsShortTimestamp, read_p_field, CcsdsTimeProvider
 from spacepackets.ecss.conf import (
     PusVersion,
     get_default_tm_apid,
@@ -25,16 +26,23 @@ from spacepackets.ecss.conf import (
 )
 
 
-def get_service_from_raw_pus_packet(raw_bytearray: bytearray) -> int:
-    """Determine the service ID from a raw packet, which can be used for packet deserialization.
+class AbstractPusTm(ABC):
+    """Generic abstraction for PUS TM packets"""
 
-    It is assumed that the user already checked that the raw bytearray contains a PUS packet and
-    only basic sanity checks will be performed.
-    :raise ValueError: If raw bytearray is too short
-    """
-    if len(raw_bytearray) < 8:
-        raise ValueError
-    return raw_bytearray[7]
+    @property
+    @abstractmethod
+    def service(self) -> int:
+        pass
+
+    @property
+    @abstractmethod
+    def subservice(self) -> int:
+        pass
+
+    @property
+    @abstractmethod
+    def source_data(self) -> bytes:
+        pass
 
 
 class PusTmSecondaryHeader:
@@ -47,7 +55,7 @@ class PusTmSecondaryHeader:
         self,
         service: int,
         subservice: int,
-        time: CdsShortTimestamp,
+        time_provider: CcsdsTimeProvider,
         message_counter: int,
         dest_id: int = 0,
         spacecraft_time_ref: int = 0,
@@ -56,7 +64,7 @@ class PusTmSecondaryHeader:
 
         :param service:
         :param subservice:
-        :param time: Time field
+        :param time_provider: Time field provider which can provide or read a time field
         :param message_counter: 8 bit counter for PUS A, 16 bit counter for PUS C
         :param dest_id: Destination ID if PUS C is used
         :param spacecraft_time_ref: Space time reference if PUS C is used
@@ -75,14 +83,14 @@ class PusTmSecondaryHeader:
             )
         self.message_counter = message_counter
         self.dest_id = dest_id
-        self.time = time
+        self.time_provider = time_provider
 
     @classmethod
     def __empty(cls) -> PusTmSecondaryHeader:
         return PusTmSecondaryHeader(
             service=0,
             subservice=0,
-            time=CdsShortTimestamp.init_from_current_time(),
+            time_provider=CdsShortTimestamp.from_current_time(),
             message_counter=0,
         )
 
@@ -93,14 +101,18 @@ class PusTmSecondaryHeader:
         secondary_header.append(self.subservice)
         secondary_header.extend(struct.pack("!H", self.message_counter))
         secondary_header.extend(struct.pack("!H", self.dest_id))
-        secondary_header.extend(self.time.pack())
+        secondary_header.extend(self.time_provider.pack())
         return secondary_header
 
     @classmethod
-    def unpack(cls, header_start: bytes) -> PusTmSecondaryHeader:
+    def unpack(
+        cls, header_start: bytes, time_reader: CcsdsTimeProvider
+    ) -> PusTmSecondaryHeader:
         """Unpack the PUS TM secondary header from the raw packet starting at the header index.
 
         :param header_start:
+        :param time_reader: Generic time reader which knows the time stamp size and how to interpret
+            the raw timestamp
         :raises ValueError: bytearray too short or PUS version missmatch.
         :return:
         """
@@ -140,17 +152,16 @@ class PusTmSecondaryHeader:
         time_code_id = read_p_field(header_start[current_idx])
         if time_code_id:
             pass
-        secondary_header.time = CdsShortTimestamp.unpack(
-            time_field=header_start[
-                current_idx : current_idx + PusTelemetry.PUS_TIMESTAMP_SIZE
-            ]
+        secondary_header.time_provider = time_reader
+        time_reader.read_from_raw(
+            header_start[current_idx : current_idx + time_reader.len]
         )
         return secondary_header
 
     def __repr__(self):
         return (
             f"{self.__class__.__name__}(service={self.service!r}, subservice={self.subservice!r}, "
-            f"time={self.time!r}, message_counter={self.message_counter!r}, "
+            f"time={self.time_provider!r}, message_counter={self.message_counter!r}, "
             f"dest_id={self.dest_id!r}, spacecraft_time_ref={self.spacecraft_time_ref!r}, "
             f"pus_version={self.pus_version!r})"
         )
@@ -160,10 +171,10 @@ class PusTmSecondaryHeader:
 
     @property
     def header_size(self) -> int:
-        return self.time.len() + 7
+        return self.time_provider.len + 7
 
 
-class PusTelemetry:
+class PusTelemetry(AbstractPusTm):
     """Generic PUS telemetry class representation.
 
     Can be used to generate TM packets using a high level interface with the default constructor,
@@ -185,11 +196,14 @@ class PusTelemetry:
     CDS_SHORT_SIZE = 7
     PUS_TIMESTAMP_SIZE = CDS_SHORT_SIZE
 
+    # TODO: Better timestamp abstraction.
+    #       Use a better default constructor which takes a parameter aggregation
+    #       Supply this constructor as a classmethod in reduced form
     def __init__(
         self,
         service: int,
         subservice: int,
-        time: Optional[CdsShortTimestamp] = None,
+        time_provider: Optional[CcsdsTimeProvider] = None,
         source_data: bytearray = bytearray([]),
         seq_count: int = 0,
         apid: int = FETCH_GLOBAL_APID,
@@ -197,23 +211,19 @@ class PusTelemetry:
         space_time_ref: int = 0b0000,
         destination_id: int = 0,
         packet_version: int = 0b000,
-        secondary_header_flag: bool = True,
     ):
         if apid == FETCH_GLOBAL_APID:
             apid = get_default_tm_apid()
-        if time is None:
-            time = CdsShortTimestamp.init_from_current_time()
-        # packet type for telemetry is 0 as specified in standard
-        # specified in standard
-        packet_type = PacketTypes.TM
+        if time_provider is None:
+            time_provider = CdsShortTimestamp.from_current_time()
         self._source_data = source_data
         data_length = self.data_len_from_src_len_timestamp_len(
-            timestamp_len=time.len(), source_data_len=len(self._source_data)
+            timestamp_len=time_provider.len, source_data_len=len(self._source_data)
         )
         self.sp_header = SpacePacketHeader(
             apid=apid,
-            packet_type=packet_type,
-            sec_header_flag=secondary_header_flag,
+            packet_type=PacketTypes.TM,
+            sec_header_flag=True,
             ccsds_version=packet_version,
             data_len=data_length,
             seq_count=seq_count,
@@ -224,7 +234,7 @@ class PusTelemetry:
             message_counter=message_counter,
             dest_id=destination_id,
             spacecraft_time_ref=space_time_ref,
-            time=time,
+            time_provider=time_provider,
         )
         self._valid = True
         self._crc16 = 0
@@ -232,7 +242,7 @@ class PusTelemetry:
     @classmethod
     def __empty(cls) -> PusTelemetry:
         return PusTelemetry(
-            service=0, subservice=0, time=CdsShortTimestamp.init_from_current_time()
+            service=0, subservice=0, time_provider=CdsShortTimestamp.from_current_time()
         )
 
     def pack(self, calc_crc: bool = True) -> bytearray:
@@ -261,15 +271,21 @@ class PusTelemetry:
         self._crc16 = crc.crcValue
 
     @classmethod
-    def unpack(cls, raw_telemetry: bytes) -> PusTelemetry:
+    def unpack(
+        cls, raw_telemetry: bytes, time_reader: Optional[CcsdsTimeProvider] = None
+    ) -> PusTelemetry:
         """Attempts to construct a generic PusTelemetry class given a raw bytearray.
-        :raises ValueError: if the format of the raw bytearray is invalid, for example the length
+
         :param raw_telemetry:
+        :param time_reader:
+        :raises ValueError: if the format of the raw bytearray is invalid, for example the length
         """
         if raw_telemetry is None:
             raise ValueError("Given byte stream invalid")
         elif len(raw_telemetry) == 0:
             raise ValueError("Given byte stream is empty")
+        if time_reader is None:
+            time_reader = CdsShortTimestamp.empty()
         pus_tm = cls.__empty()
         pus_tm._valid = False
         pus_tm.sp_header = SpacePacketHeader.unpack(space_packet_raw=raw_telemetry)
@@ -284,7 +300,8 @@ class PusTelemetry:
             )
             raise ValueError
         pus_tm.pus_tm_sec_header = PusTmSecondaryHeader.unpack(
-            header_start=raw_telemetry[SPACE_PACKET_HEADER_SIZE:]
+            header_start=raw_telemetry[SPACE_PACKET_HEADER_SIZE:],
+            time_reader=time_reader,
         )
         if (
             expected_packet_len
@@ -307,6 +324,18 @@ class PusTelemetry:
         )[0]
         pus_tm.__perform_crc_check(raw_telemetry=raw_telemetry[:expected_packet_len])
         return pus_tm
+
+    @staticmethod
+    def service_from_bytes(raw_bytearray: bytearray) -> int:
+        """Determine the service ID from a raw packet, which can be used for packet deserialization.
+
+        It is assumed that the user already checked that the raw bytearray contains a PUS packet and
+        only basic sanity checks will be performed.
+        :raise ValueError: If raw bytearray is too short
+        """
+        if len(raw_bytearray) < 8:
+            raise ValueError
+        return raw_bytearray[7]
 
     @classmethod
     def from_composite_fields(
@@ -369,13 +398,17 @@ class PusTelemetry:
         return self.pus_tm_sec_header.subservice
 
     @property
+    def source_data(self) -> bytes:
+        return self.tm_data
+
+    @property
     def valid(self) -> bool:
         return self._valid
 
     @property
     def tm_data(self) -> bytearray:
         """
-        :return: TM application data (raw)
+        :return: TM source data (raw)
         """
         return self._source_data
 
@@ -383,8 +416,24 @@ class PusTelemetry:
     def tm_data(self, data: bytes):
         self._source_data = data
         self.sp_header.data_len = self.data_len_from_src_len_timestamp_len(
-            self.pus_tm_sec_header.time.len(), len(data)
+            self.pus_tm_sec_header.time_provider.len, len(data)
         )
+
+    @property
+    def apid(self):
+        return self.sp_header.apid
+
+    @apid.setter
+    def apid(self, apid: int):
+        self.sp_header.apid = apid
+
+    @property
+    def seq_flags(self):
+        return self.sp_header.seq_flags
+
+    @seq_flags.setter
+    def seq_flags(self, seq_flags):
+        self.sp_header.seq_flags = seq_flags
 
     @property
     def packet_id(self):
@@ -427,10 +476,6 @@ class PusTelemetry:
         the space packet header.
         """
         return self.sp_header.packet_len
-
-    @property
-    def apid(self) -> int:
-        return self.sp_header.apid
 
     @property
     def seq_count(self) -> int: