+++ tests/ecss/test_pus_tm.py
@@ -1,6 +1,10 @@
 #!/usr/bin/env python3
+import struct
 from typing import Optional
 from unittest import TestCase
+from unittest.mock import MagicMock
+
+from crcmod.predefined import mkPredefinedCrcFun
 
 from spacepackets.ccsds.spacepacket import (
     PacketId,
@@ -9,6 +13,7 @@ from spacepackets.ccsds.spacepacket import (
     SequenceFlags,
     SpacePacketHeader,
 )
+from spacepackets.ccsds.time import CcsdsTimeProvider
 from spacepackets.ecss import PacketFieldEnum
 from spacepackets.ecss.tc import PusTelecommand
 from spacepackets.ecss.conf import set_default_tm_apid
@@ -17,7 +22,6 @@ from spacepackets.ecss.tm import (
     PusTelemetry,
     CdsShortTimestamp,
     PusVersion,
-    get_service_from_raw_pus_packet,
     PusTmSecondaryHeader,
 )
 from spacepackets.ecss.pus_17_test import Service17Tm
@@ -42,138 +46,214 @@ from spacepackets.ecss.pus_1_verification import (
 
 
 class TestTelemetry(TestCase):
-    def test_telemetry(self):
-        pus_17_tm = PusTelemetry(
+    def setUp(self) -> None:
+        self.time_stamp_provider = MagicMock(spec=CcsdsTimeProvider)
+        self.time_stamp_provider.len = 7
+        self.raw_stamp = bytes([0, 1, 2, 3, 4, 5, 6])
+        self.time_stamp_provider.pack.return_value = self.raw_stamp
+        self.ping_reply = PusTelemetry(
             service=17,
             subservice=2,
-            apid=0xEF,
-            seq_count=22,
+            apid=0x123,
+            seq_count=0x234,
             source_data=bytearray(),
-            time=CdsShortTimestamp.init_from_current_time(),
+            time_provider=self.time_stamp_provider,
         )
-        self.assertEqual(pus_17_tm.get_source_data_string(PrintFormats.HEX), "hex []")
-        self.assertEqual(pus_17_tm.get_source_data_string(PrintFormats.DEC), "dec []")
-        self.assertEqual(pus_17_tm.get_source_data_string(PrintFormats.BIN), "bin []")
-        self.assertEqual(pus_17_tm.subservice, 2)
-        self.assertEqual(pus_17_tm.service, 17)
-        self.assertEqual(pus_17_tm.seq_count, 22)
-        self.assertEqual(pus_17_tm.packet_len, 22)
-        pus_17_raw = pus_17_tm.pack()
-        self.assertEqual(get_service_from_raw_pus_packet(raw_bytearray=pus_17_raw), 17)
-        self.assertRaises(ValueError, get_service_from_raw_pus_packet, bytearray())
+        self.ping_reply_raw = self.ping_reply.pack()
 
+    def test_state(self):
+        self.assertEqual(
+            self.ping_reply.get_source_data_string(PrintFormats.HEX), "hex []"
+        )
+        self.assertEqual(
+            self.ping_reply.get_source_data_string(PrintFormats.DEC), "dec []"
+        )
+        self.assertEqual(
+            self.ping_reply.get_source_data_string(PrintFormats.BIN), "bin []"
+        )
+        self.assertEqual(self.ping_reply.subservice, 2)
+        self.assertEqual(self.ping_reply.service, 17)
+        self.assertEqual(self.ping_reply.apid, 0x123)
+        self.assertEqual(self.ping_reply.seq_count, 0x234)
+        self.assertEqual(self.ping_reply.sp_header.data_len, 15)
+        self.assertEqual(self.ping_reply.packet_len, 22)
+
+    def test_raw(self):
+        # Secondary header is set -> 0b0000_1001 , APID occupies last bit of first byte
+        self.assertEqual(self.ping_reply_raw[0], 0x09)
+        # Rest of APID
+        self.assertEqual(self.ping_reply_raw[1], 0x23)
+        # Unsegmented is the default, and first byte of 0x234 occupies this byte as well
+        self.assertEqual(self.ping_reply_raw[2], 0xC2)
+        self.assertEqual(self.ping_reply_raw[3], 0x34)
+        self.assertEqual((self.ping_reply_raw[4] << 8) | self.ping_reply_raw[5], 15)
+        # SC time ref status is 0
+        self.assertEqual(self.ping_reply_raw[6], PusVersion.PUS_C << 4)
+        self.assertEqual(self.ping_reply_raw[7], 17)
+        self.assertEqual(self.ping_reply_raw[8], 2)
+        # MSG counter
+        self.assertEqual(self.ping_reply_raw[9], 0x00)
+        self.assertEqual(self.ping_reply_raw[10], 0x00)
+        # Destination ID
+        self.assertEqual(self.ping_reply_raw[11], 0x00)
+        self.assertEqual(self.ping_reply_raw[12], 0x00)
+        self.assertEqual(self.ping_reply_raw[13 : 13 + 7], self.raw_stamp)
+        # CRC16-CCITT checksum
+        crc_func = mkPredefinedCrcFun(crc_name="crc-ccitt-false")
+        data_to_check = self.ping_reply_raw[0:20]
+        crc16 = crc_func(data_to_check)
+        self.assertEqual(crc16, struct.unpack("!H", self.ping_reply_raw[20:22])[0])
+
+    def test_state_setting(self):
+        self.ping_reply.sp_header.apid = 0x22
+        source_data = bytearray([0x42, 0x38])
+        self.ping_reply.tm_data = source_data
+        self.assertEqual(self.ping_reply.apid, 0x22)
+        self.assertEqual(
+            self.ping_reply.pus_tm_sec_header.pus_version, PusVersion.PUS_C
+        )
+        self.assertTrue(self.ping_reply.valid)
+        self.assertEqual(self.ping_reply.tm_data, source_data)
+        self.assertEqual(self.ping_reply.packet_id.raw(), 0x0822)
+        self.assertEqual(self.ping_reply.packet_len, 24)
+
+    def test_service_from_raw(self):
+        self.assertEqual(
+            PusTelemetry.service_from_bytes(raw_bytearray=self.ping_reply_raw), 17
+        )
+
+    def test_service_from_raw_invalid(self):
+        self.assertRaises(ValueError, PusTelemetry.service_from_bytes, bytearray())
+
+    def test_source_data_string_getters(self):
         set_default_tm_apid(0x22)
         source_data = bytearray([0x42, 0x38])
-        pus_17_tm = PusTelemetry(
-            service=17,
-            subservice=2,
-            seq_count=22,
-            source_data=source_data,
+        self.ping_reply.tm_data = source_data
+        self.assertEqual(
+            self.ping_reply.get_source_data_string(PrintFormats.HEX), "hex [42,38]"
         )
         self.assertEqual(
-            pus_17_tm.get_source_data_string(PrintFormats.HEX), "hex [42,38]"
+            self.ping_reply.get_source_data_string(PrintFormats.DEC), "dec [66,56]"
         )
         self.assertEqual(
-            pus_17_tm.get_source_data_string(PrintFormats.DEC), "dec [66,56]"
-        )
-        self.assertEqual(
-            pus_17_tm.get_source_data_string(PrintFormats.BIN),
+            self.ping_reply.get_source_data_string(PrintFormats.BIN),
             "bin [\n0:01000010\n1:00111000\n]",
         )
-        self.assertEqual(pus_17_tm.apid, 0x22)
-        self.assertEqual(pus_17_tm.pus_tm_sec_header.pus_version, PusVersion.PUS_C)
-        self.assertTrue(pus_17_tm.valid)
-        self.assertEqual(pus_17_tm.tm_data, source_data)
-        self.assertEqual(pus_17_tm.packet_id.raw(), 0x0822)
-        pus_17_tm.print_full_packet_string(PrintFormats.HEX)
-        self.assertEqual(pus_17_tm.packet_len, 24)
-        crc16 = pus_17_tm.crc16
-        crc_string = f"{(crc16 & 0xff00) >> 8:02x},{crc16 & 0xff:02x}"
-        raw_time = pus_17_tm.pus_tm_sec_header.time.pack()
-        raw_space_packet_header = pus_17_tm.sp_header.pack()
-        sp_header_as_str = raw_space_packet_header.hex(sep=",", bytes_per_sep=1)
-        raw_secondary_packet_header = pus_17_tm.pus_tm_sec_header.pack()
+
+    def test_print(self):
+        self.ping_reply.print_full_packet_string(PrintFormats.HEX)
+
+    def test_sec_header(self):
+        raw_secondary_packet_header = self.ping_reply.pus_tm_sec_header.pack()
         self.assertEqual(raw_secondary_packet_header[0], 0x20)
         # Service
         self.assertEqual(raw_secondary_packet_header[1], 17)
         # Subservice
         self.assertEqual(raw_secondary_packet_header[2], 2)
+
+    def test_full_printout(self):
+        crc16 = self.ping_reply.crc16
+        crc_string = f"{(crc16 & 0xff00) >> 8:02x},{crc16 & 0xff:02x}"
+        raw_time = self.ping_reply.pus_tm_sec_header.time_provider.pack()
+        raw_space_packet_header = self.ping_reply.sp_header.pack()
+        sp_header_as_str = raw_space_packet_header.hex(sep=",", bytes_per_sep=1)
+        raw_secondary_packet_header = self.ping_reply.pus_tm_sec_header.pack()
         second_header_as_str = raw_secondary_packet_header.hex(sep=",", bytes_per_sep=1)
         expected_printout = f"hex [{sp_header_as_str},{second_header_as_str},"
-        expected_printout += "42,38,"
         expected_printout += f"{crc_string}]"
         self.assertEqual(
-            pus_17_tm.get_full_packet_string(PrintFormats.HEX), expected_printout
+            self.ping_reply.get_full_packet_string(PrintFormats.HEX), expected_printout
         )
-        pus_17_raw = pus_17_tm.pack()
 
-        pus_17_tm_unpacked = PusTelemetry.unpack(raw_telemetry=pus_17_raw)
-        print(pus_17_tm)
-        print(pus_17_tm.__repr__())
+    def test_print_2(self):
+        print(self.ping_reply)
+        print(self.ping_reply.__repr__())
+
+    def test_unpack(self):
+        source_data = bytearray([0x42, 0x38])
+        self.ping_reply.tm_data = source_data
+        self.ping_reply.sp_header.apid = 0x22
+        self.ping_reply_raw = self.ping_reply.pack()
+        self.time_stamp_provider.read_from_raw = MagicMock()
+        pus_17_tm_unpacked = PusTelemetry.unpack(
+            raw_telemetry=self.ping_reply_raw, time_reader=self.time_stamp_provider
+        )
+
         self.assertEqual(pus_17_tm_unpacked.apid, 0x22)
         self.assertEqual(
             pus_17_tm_unpacked.pus_tm_sec_header.pus_version, PusVersion.PUS_C
         )
         self.assertTrue(pus_17_tm_unpacked.valid)
+        pus_17_tm_unpacked.pus_tm_sec_header.time_provider.read_from_raw.assert_called_once()
+        pus_17_tm_unpacked.pus_tm_sec_header.time_provider.read_from_raw.assert_called_with(
+            self.raw_stamp
+        )
         self.assertEqual(pus_17_tm_unpacked.tm_data, source_data)
         self.assertEqual(pus_17_tm_unpacked.packet_id.raw(), 0x0822)
-        self.assertRaises(ValueError, PusTelemetry.unpack, None)
-        self.assertRaises(ValueError, PusTelemetry.unpack, bytearray())
-        correct_size = pus_17_raw[4] << 8 | pus_17_raw[5]
-        # Set length field invalid
-        pus_17_raw[4] = 0x00
-        pus_17_raw[5] = 0x00
-        self.assertRaises(ValueError, PusTelemetry.unpack, pus_17_raw)
-        pus_17_raw[4] = 0xFF
-        pus_17_raw[5] = 0xFF
-        self.assertRaises(ValueError, PusTelemetry.unpack, pus_17_raw)
 
-        pus_17_raw[4] = (correct_size & 0xFF00) >> 8
-        pus_17_raw[5] = correct_size & 0xFF
-        pus_17_raw.append(0)
+        correct_size = self.ping_reply_raw[4] << 8 | self.ping_reply_raw[5]
+        # Set length field invalid
+        self.ping_reply_raw[4] = 0x00
+        self.ping_reply_raw[5] = 0x00
+        self.assertRaises(ValueError, PusTelemetry.unpack, self.ping_reply_raw)
+        self.ping_reply_raw[4] = 0xFF
+        self.ping_reply_raw[5] = 0xFF
+        self.assertRaises(ValueError, PusTelemetry.unpack, self.ping_reply_raw)
+
+        self.ping_reply_raw[4] = (correct_size & 0xFF00) >> 8
+        self.ping_reply_raw[5] = correct_size & 0xFF
+        self.ping_reply_raw.append(0)
         # Should work with a warning
-        pus_17_tm_unpacked = PusTelemetry.unpack(raw_telemetry=pus_17_raw)
+        pus_17_tm_unpacked = PusTelemetry.unpack(
+            raw_telemetry=self.ping_reply_raw, time_reader=self.time_stamp_provider
+        )
 
         # This should cause the CRC calculation to fail
         incorrect_size = correct_size + 1
-        pus_17_raw[4] = (incorrect_size & 0xFF00) >> 8
-        pus_17_raw[5] = incorrect_size & 0xFF
-        pus_17_tm_unpacked = PusTelemetry.unpack(raw_telemetry=pus_17_raw)
+        self.ping_reply_raw[4] = (incorrect_size & 0xFF00) >> 8
+        self.ping_reply_raw[5] = incorrect_size & 0xFF
+        pus_17_tm_unpacked = PusTelemetry.unpack(
+            raw_telemetry=self.ping_reply_raw, time_reader=self.time_stamp_provider
+        )
 
+    def test_faulty_unpack(self):
+        self.assertRaises(ValueError, PusTelemetry.unpack, None, None)
+        self.assertRaises(ValueError, PusTelemetry.unpack, bytearray(), None)
+
+    def test_invalid_sec_header_unpack(self):
         invalid_secondary_header = bytearray([0x20, 0x00, 0x01, 0x06])
         self.assertRaises(
-            ValueError, PusTmSecondaryHeader.unpack, invalid_secondary_header
+            ValueError, PusTmSecondaryHeader.unpack, invalid_secondary_header, None
         )
+
+    def test_space_packet_conversion(self):
+        ccsds_packet = self.ping_reply.to_space_packet()
+        self.assertEqual(ccsds_packet.apid, self.ping_reply.apid)
+        self.assertEqual(ccsds_packet.pack(), self.ping_reply.pack())
+        self.assertEqual(ccsds_packet.seq_count, self.ping_reply.seq_count)
+        self.assertEqual(ccsds_packet.sec_header_flag, True)
+        pus_17_from_composite_fields = PusTelemetry.from_composite_fields(
+            sp_header=self.ping_reply.sp_header,
+            sec_header=self.ping_reply.pus_tm_sec_header,
+            tm_data=self.ping_reply.tm_data,
+        )
+        self.assertEqual(pus_17_from_composite_fields.pack(), self.ping_reply.pack())
+
+    def test_faulty_sec_header(self):
         with self.assertRaises(ValueError):
             # Message Counter too large
             PusTmSecondaryHeader(
                 service=0,
                 subservice=0,
-                time=CdsShortTimestamp.init_from_current_time(),
+                time_provider=CdsShortTimestamp.from_current_time(),
                 message_counter=129302,
             )
-        valid_secondary_header = PusTmSecondaryHeader(
-            service=0,
-            subservice=0,
-            time=CdsShortTimestamp.init_from_current_time(),
-            message_counter=22,
-        )
 
-        ccsds_packet = pus_17_tm.to_space_packet()
-        self.assertEqual(ccsds_packet.apid, pus_17_tm.apid)
-        self.assertEqual(ccsds_packet.pack(), pus_17_tm.pack())
-        self.assertEqual(ccsds_packet.seq_count, pus_17_tm.seq_count)
-        self.assertEqual(ccsds_packet.sec_header_flag, True)
-        pus_17_from_composite_fields = PusTelemetry.from_composite_fields(
-            sp_header=pus_17_tm.sp_header,
-            sec_header=pus_17_tm.pus_tm_sec_header,
-            tm_data=pus_17_tm.tm_data,
-        )
-        print(pus_17_from_composite_fields.pack().hex(sep=","))
-        print(pus_17_tm.pack().hex(sep=","))
-        self.assertEqual(pus_17_from_composite_fields.pack(), pus_17_tm.pack())
-        # Hand checked to see if all __repr__ were implemented properly
-        print(f"{pus_17_tm!r}")
+    def test_invalid_raw_size(self):
+        # Set length field invalid
+        self.ping_reply_raw[4] = 0x00
+        self.ping_reply_raw[5] = 0x00
+        self.assertRaises(ValueError, PusTelemetry.unpack, self.ping_reply_raw, None)
 
     def test_service_17_tm(self):
         srv_17_tm = Service17Tm(subservice=2)